feat(scripts): refactor package scope utilities to separate module
Move package scope change logic to separate module for better organization and
reuse.

Issue: BTC-1826
